Extract the router needed components into dedicated files.
