+++ src/pledger-io.tsx
@@ -1,23 +1,6 @@
-import {
-    createBrowserRouter,
-    isRouteErrorResponse,
-    LoaderFunctionArgs,
-    Outlet,
-    redirect,
-    RouterProvider,
-    useNavigate,
-    useRouteError
-} from "react-router";
-import { lazy, Suspense } from "react";
-
-import SecurityRepository from "./core/repositories/security-repository";
-import RestAPI from "./core/repositories/rest-api";
-import { CurrencyRepository } from "./core/RestAPI";
-
-import Sidebar from "./components/sidebar";
-import MobileSidebar from "./components/sidebar/mobile-sidebar";
+import {createBrowserRouter, redirect, RouterProvider} from "react-router";
+import {lazy} from "react";
 import Loading from "./components/layout/loading.component";
-import NotificationCenter from "./components/notification";
 
 import LoginPage from "./pages/login";
 import RegisterPage from "./pages/register";
@@ -33,6 +16,11 @@ import reports from "./pages/reports/routes";
 import transactions from "./pages/transaction/routes";
 import settings from "./pages/setting/routes";
 import upload from "./pages/upload/routes";
+
+import {anonymousLoader} from "./router/anonymous.loader";
+import {authenticatedLoader} from "./router/authenticated.loader";
+import {RootErrorBoundary} from "./router/error-boundary";
+import {AuthenticatedComponent} from "./router/authenticate.wrapper";
 
 const router = createBrowserRouter([
     {
@@ -57,7 +45,7 @@ const router = createBrowserRouter([
         path: '/',
         Component: AuthenticatedComponent,
         loader: authenticatedLoader,
-        errorElement: <RootErrorBoundary/>,
+        errorElement: <RootErrorBoundary />,
         hydrateFallbackElement: <Loading />,
         children: [
             {
@@ -85,74 +73,6 @@ const router = createBrowserRouter([
 ], {
     basename: '/ui'
 })
-
-// todo the functions below should be moved to a separate file
-
-function anonymousLoader({ request }: LoaderFunctionArgs) {
-    const [from] = new URLSearchParams(request.url).get("from") || "/dashboard"
-    if (SecurityRepository.isAuthenticated()) {
-        return redirect(from);
-    }
-    return null;
-}
-
-async function authenticatedLoader({ request }: LoaderFunctionArgs) {
-    const params = new URLSearchParams();
-    params.set("from", new URL(request.url).pathname.replace("/ui", ""));
-    if (!SecurityRepository.isAuthenticated()) {
-        return redirect("/login?" + params.toString());
-    }
-
-    await RestAPI.profile()
-    const profile = RestAPI.user() as any
-    await CurrencyRepository.list()
-    profile.defaultCurrency = CurrencyRepository.cached(profile.currency)
-    return {
-        user: profile
-    }
-}
-
-function AuthenticatedComponent() {
-    const logout = () => {
-        SecurityRepository.logout()
-    }
-
-    return <>
-        <Sidebar logoutCallback={ logout }/>
-        <MobileSidebar logoutCallback={ logout }/>
-        <main className='Main md:px-2 md:px-5 h-[100vh] flex flex-col overflow-y-auto'>
-            <NotificationCenter/>
-            <Suspense fallback={ <SuspenseLoading/> }>
-                <Outlet/>
-            </Suspense>
-        </main>
-    </>
-}
-
-function RootErrorBoundary() {
-    const error = useRouteError()
-    const navigate = useNavigate()
-
-    console.log('error', error)
-
-    if (!isRouteErrorResponse(error) && (error as any)?.response?.status === 403) {
-        // todo this does not work. Need to figure out a way to do a proper route redirect
-        navigate('/two-factor')
-        return <></>
-    } else if (!isRouteErrorResponse(error) && (error as any)?.response?.status === 401) {
-        SecurityRepository.logout()
-        window.location.reload()
-        return <></>
-    }
-
-    return <></>
-}
-
-function SuspenseLoading() {
-    return <div className='flex h-[100vh] justify-center items-center'>
-        <Loading />
-    </div>
-}
 
 function _() {
     return (
